Moved server routes to their own folder.
