+++ client/src/components/pet-components/EditPetForm.jsx
@@ -46,7 +46,7 @@ export function EditPetForm({ baseURL }) {
 
   async function loadPet(id) {
     try {
-      await fetch(`https://${baseURL}/pet?id=${id}`)
+      await fetch(`https://${baseURL}/petbyid?id=${id}`)
         .then((res) => {
           if (res.status >= 400) {
             throw res.status;
